moved client only methods to client project
